Add missing DTOs (UserDto, ChatDto), fix CommentService Long type, remove FRIENDS_ONLY visibility, fix AdminUserService to use entity collections

+++ spring_boot_backend/src/main/kotlin/com/androidinsta/Service/CommentService.kt
@@ -123,10 +123,10 @@ class CommentService(
         return count
     }
 
-    fun getRepliesCount(commentId: Long): Int {
+    fun getRepliesCount(commentId: Long): Long {
         val cacheKey = "replies:count:$commentId"
         
-        val cached = redisService.get(cacheKey, Int::class.java)
+        val cached = redisService.get(cacheKey, Long::class.java)
         if (cached != null) {
             return cached
         }
